[Distributed] Introduce DistributedDecl.h for greater reuse
